Add basic initial unit & integration tests
There's fairly limited scope for testing internally in this package, as the majority of
it is dependent on being a single process (e.g. we deliberately clear all output buffers
and fail if headers are already sent, so can't execute more than one
MicroFramework->execute() in the same process). So the majority of testing will need to
be done as a black box test.