Add blackbox coverage of logging
Cover the key functionality of the stackdriver logger, including request
logging, exception logging, and custom logging. I've tested with this
because it's the only logger microframework directly supports, and because
it's complex enough that we should be able to have confidence any generic
PSR\LoggerInterface would work as expected if wired up correctly.