Only build with current versions of dependencies
The `--lowest` build won't apply as-is to the blackbox tests anyway,
and would require a bit of fiddly wiring to carry that config over.
For this package, the dependencies are limited and will be kept
fairly tight, so there's limited value in the extra complexity of
running a separate --lowest build.